-- Removed 'vertical_cursor' code.  You can still get a vertical cursor by calling curs_set( 5).
-- Removed some debug printf()s for combined characters that never
ought to have been left in.

-- The cursor flickered during frequent screen redraws or cursor
movements.  In _display_screen() and _refresh_screen(),  the cursor is
now redrawn if the line it's on has been redrawn.

-- The cursor is supposed to change to an outlined rectangle if you're
in "compose" mode.  This has only worked if you aren't using XIM.  (I
tried to get it to work when using XIM,  but I don't see how you detect
that we're in "compose" mode.  It would be nice if it worked,  since
XIM is the recommended configuration;  see README.md.)

   Also,  the cursor is supposed to stop blinking when the cursor leaves
the window.  The logic for both of these effects is considerably simpler
now than it was.

-- There were a lot of places where _redraw_cursor( ) was called that
could be eliminated,  now that we know it'll be updated every half second
anyway.

-- Cursor types 'invisible',  'normal',  'block',  (outlined) 'rectangle',
'vertical',  and 'half block' are now supported,  and you can blink between
any two of these,  as with the Win32a flavor of PDCurses.