'vcwin32.mak' is now the makefile for both MS Visual C/C++ and for the Intel(R) compiler.  For the latter,  add ICC=Y (only difference is use of icl.exe and xilink.exe instead of cl.exe and link.exe,  so having two separate makefiles didn't really seem reasonable).  As a result, iccwin32.mak has been removed.
Also,  add IX86=Y to do a 32-bit build instead of the default 64-bit.
(Previously,  you had to edit the makefile to accomplish this;
Nadav-Eyesight pointed out that this was not a good idea.)

This applies,  thus far,  to the Win32a and console Win32 flavors.
(Note that the various '32' names for flavors and makefiles are
increasingly misleading.)  The vcwin32.mak for the console flavor
now also supports switches to choose 32-bit or 16-bit chtypes,
making it look very much like the Win32a version of vcwin32.mak.