SDL2: Use SDL_BlitSurface() instead of the semi-private SDL_LowerBlit() so that rect validation and clipping happen when we draw
While using SDL_LowerBlit() I get a segfault from the SDL blitting code
where an out-of-bounds access is taking place when SDL tries to blit,
running off the edge of the surface due to the rect being wrong for the
current surface size, and/or the fact that the usual clipping isn't
taking place.

This happens easily with a tiling WM when you allow for
SDL_WINDOW_RESIZABLE.

After making this change, I found this change had already been proposed
by @jpcima in PR #35.