Fix non-wide sdl2 font coloration
With SDL 2.0.14, it seems like the palette size of a loaded bitmap
surface is larger than the palette size of the actual bitmap.
Therefore, the last color in the palette never actually shows up in
the font. This leads to characters not being properly colored.

This patch adds some complex code because it needs to parse part of
the bitmap itself. A much simpler way to fix this issue would be to
use the first and second palette entries rather than the first and
last, but doing so would break custom fonts with more than two
colors.