Handle KEY_BACKSPACE in tui.c for tuidemo
Due to the erasechar() API, which returns char, and wgetch(), which
returns int, and KEY_BACKSPACE, which is int, it is incorrect
to completely rely on erasechar() for backspace support.
If some systems truly return KEY_BACKSPACE, then checking against
erasechar() alone will never catch the true return value.

This is primarily a compatibility fix for people like me who test
the demos with other curses libraries.  It worked fine in PDCurses,
but not ncurses.